requirements.txt
crawl4ai==0.7.1
loguru==0.7.0
tiktoken==1.39.0
pydantic==2.10.2
pydantic-settings==2.0.0
django==5.0.1
django-cors-headers==4.0.0
django-rest-framework==0.1.0
datasets==2.14.5
langchain==0.2.11
langchain-community==0.2.11
langchain-huggingface==0.2.11
zenml==0.39.0
boto3==1.36.0
langchain-openai==0.3.0
mlflow==2.1.1